Extract CreateObjectService from CocinaObjectStore
Because CocinaObjectStore was too long and had too many responsibilities
